Rename the legacy `openssl` package to `openssl1`
OpenSSL v1.1 is maintained until September 2023, after that it will no
longer receive security updates.

MSYS2 already switched to OpenSSL v3, but Git for Windows is not quite
yet prepared for that switch.

This causes problems e.g. when trying to run `action-tmate` in a Git for
Windows SDK because only Git for Windows' packages expect the OpenSSL
v1.1 libraries to be present, but MSYS2's packages expect OpenSSL v3 to
be present. Since the version is encoded in the `.dll` files' names, it
totally breaks one part of the equation.

To work around that, let's use a different package name altogether for
OpenSSL v1.1 and rebuild all the dependencees.

Note: We do not use the name `mingw-w64-openssl-1.1` because there is
Git for Windows tooling that wants to detect the version in a package
name by looking for the pattern `-[0-9]`, which would get confused.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>