git: we're actually not depending on `openssl`
The `git` executable has no dependency on OpenSSL. But it was still
listed as such, a left-over from a time long past (the last time Git for
Windows depended on OpenSSL was before switching to SHA-1DC).

Note: The `git-remote-https` executable depends on cURL, which can
itself be linked to OpenSSL, but that's for the `mingw-w64-curl`
dependencies to specify, not for `mingw-w64-git`'s.

Also note: The only (and quite obscure) part of Git for Windows that
(inadvertently) depended directly on OpenSSL was `git-imap-send`. It is
unclear how many Git for Windows users rely on that command (a couple of
very vocal voices prevent the Git for Windows project from adding even
just opt-in telemetry that would equip the project with invaluable data
to answer questions like this), but on the off chance that there _are_
users, let's just let libcurl handle the IMAP stuff instead, as we
depend on libcurl already for fetching/pushing via HTTPS _anyway_.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>